requirements.txt
Flask~=3.0.0
flask_sqlalchemy==3.0.5
python-dotenv~=1.0.0
Flask_WTF==1.1.1
WTForms==3.0.1
Bootstrap_Flask==2.2.0
Werkzeug==2.2.3
wtforms[email]==2.3.1
pandas~=2.1.3